api_rest/requirements.txt
flask==2.0.2
minio===7.1.0
pyarrow==4.0.1
pandas==1.3.3
flask-restful==0.3.9
psycopg2-binary==2.9.1